Minor refactoring
Part of #21

Extracted the StagingTableProcessor class
